manager tests - inputs

+++ view/input_manager.py
@@ -96,7 +96,7 @@ class InputManager:
     
     def updating_product_price(self):
         raw_input = input().split(' ')
-        price = int(raw_input[-1])
+        price = float(raw_input[-1])
         product = ' '.join(raw_input[:-1])
         return product, price
